==> 1_Data_Collection_and_Preparation/4_Fill_Empty_Genre.ipynb <==
## Fill Empty Genres

# import libraries
import pandas as pd

# access file 
file_path = '../Cleaned_Data/Artist_Genre_List.csv'
artist_genre_df = pd.read_csv(file_path)
filter_genre_df = artist_genre_df.copy()

### Filter rows with valid genre data

# drop rows where genre is 'no genres found'
filter_genre_df = artist_genre_df[artist_genre_df['genres'] != 'no genres found']


# reset index
filter_genre_df = filter_genre_df.reset_index(drop=True)
print(len(filter_genre_df))

### Add filled genre data

# append filled genres
# file paths
genres_filled_file_path = '../Cleaned_Data/Empty_Genres_Filled.csv'
# load dataframes
genre_filled_df = pd.read_csv(genres_filled_file_path)

# combine both DataFrames
combined_df = pd.concat([filter_genre_df,genre_filled_df], ignore_index=True)

# remove duplicates based on artist_name and genres
combined_df=combined_df.drop_duplicates(subset=['artist_name','genres']).reset_index(drop=True)




### Save as new file

output_path = '../Cleaned_Data/Completed_Artist_Genre_List.csv'
combined_df.to_csv(output_path, index=False, encoding='utf-8-sig')
